feat: add support for OpenAPI server
