+++ Cargo.toml
@@ -70,6 +70,8 @@ tracing = "0.1"
 tracing-subscriber = "0.3"
 tugraph = "0.1.4"
 url = "2.5"
+utoipa = "5.3"
+utoipa-swagger-ui = "9.0"
 uuid = "1.12"
 walkdir = "2.5"
 zip = "2.2"
