+++ deploy-crates-pro.sh
@@ -9,7 +9,7 @@ NAMESPACE=crates-pro
 INSTANCE=test1
 DEPLOYMENT=cratespro-backend-$INSTANCE
 KAFKA_HOST=172.17.0.1:30092
-TAKE_SNAPSHOT_BEFORE_REDEPLOY=1
+TAKE_SNAPSHOT_BEFORE_REDEPLOY=0
 # build
 TIMESTAMP=$(date +%Y%m%d-%H%M)
 STAGE1_IMAGE=crates-pro-infra:base-$TIMESTAMP
